# openapi.yml
openapi: 3.0.3
info:
  title: Luga Store API
  version: 1.0.0
  description: API documentation for the Luga Store backend, listing all customer, shop manager/owner, and administrator endpoints.
security:
  - CookieAuth: []
paths:
...
  /categories/{id}:
    put:
      tags: [Categories (Global)]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryRequest'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags: [Categories (Global)]
      summary: Delete a category (high performance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
      responses:
      '200':
        description: Category deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /categories/{id}/status:
    patch:
      tags: [Categories (Global)]
      summary: Toggle category active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryActiveStatusRequest'
      responses:
        '200':
          description: Active status updated successfully
  /categories/reorder:
    post:
      tags: [Categories (Global)]
      summary: Reorder categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        '200':
          description: Reordered successfully

  # --- SHOP-LEVEL CATEGORIES ---
  /shops/{shopId}/categories:
    get:
      tags: [Categories (Shop)]
      summary: Get all categories of a shop
      security: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: name
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Categories list retrieved

    post:
      tags: [Categories (Shop)]
      summary: Create a category under a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopUpsertCategoryRequest'
      responses:
        '200':
          description: Category created successfully
  /shops/{shopId}/categories/{id}:
    put:
      tags: [Categories (Shop)]
      summary: Update a category under a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopUpsertCategoryRequest'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags: [Categories (Shop)]
      summary: Delete a shop category
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Category deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{shopId}/categories/{id}/status:
    patch:
      tags: [Categories (Shop)]
      summary: Toggle active status of a shop category
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryActiveStatusRequest'
      responses:
        '200':
          description: Status toggled successfully
  /shops/{shopId}/categories/reorder:
    post:
      tags: [Categories (Shop)]
      summary: Reorder categories of a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        '200':
          description: Reordered successfully

  # --- GLOBAL PRODUCTS ---
...
  /shops/{id}:
    get:
      tags: [Shops]
      summary: Get details of a single shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop details

    put:
      tags: [Shops]
      summary: Update a shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertShopRequest'
      responses:
        '200':
          description: Shop updated successfully

    delete:
      tags: [Shops]
      summary: Delete a shop (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Shop deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{id}/status:
    patch:
      tags: [Shops]
      summary: Toggle active status of a shop (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopActiveStatusRequest'
      responses:
        '200':
          description: Shop active status updated successfully

  # --- SHOP OWNERS ---
...
  /shops/{shopId}/managers:
    get:
      tags: [Shop Members (Managers)]
      summary: Get all managers for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of shop managers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'

    post:
      tags: [Shop Members (Managers)]
      summary: Invite a new manager to a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopMemberInviteRequest'
      responses:
        '200':
          description: Manager invited successfully
  /shops/{shopId}/managers/invited:
    get:
      tags: [Shop Members (Managers)]
      summary: Get paginated list of invited managers for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of invited shop managers retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvitedUserList'
  /shops/{shopId}/managers/{managerId}:
    get:
      tags: [Shop Members (Managers)]
      summary: Get details of a specific shop manager
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: managerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Manager details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    patch:
      tags: [Shop Members (Managers)]
      summary: Update shop manager details
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: managerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopMemberRequest'
      responses:
        '200':
          description: Manager updated successfully

    delete:
      tags: [Shop Members (Managers)]
      summary: Delete a shop manager
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: managerId
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Manager deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{shopId}/managers/{managerId}/resend-invitation:
    post:
      tags: [Shop Members (Managers)]
      summary: Resend shop manager invitation
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: managerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invitation resent successfully
  /shops/{shopId}/managers/{managerId}/invitation:
    delete:
      tags: [Shop Members (Managers)]
      summary: Delete pending shop manager invitation
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: managerId
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Invitation deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string

  # --- DASHBOARD SUMMARIES (Planned) ---
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
...
  schemas:
...
    UpsertCategoryRequest:
      type: object
      required: [name]
      properties:
        shopId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
    ShopUpsertCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    UpdateCategoryActiveStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
        shopId:
          type: integer
          nullable: true
    ReorderCategoriesRequest:
      type: object
      required: [orders]
      properties:
        shopId:
          type: integer
          nullable: true
        orders:
          type: array
          items:
            type: object
            required: [id, displayOrder]
            properties:
              id:
                type: integer
              displayOrder:
                type: integer
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        avatarFileName:
          type: string
          nullable: true
        isEmailConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InvitedUserDto:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    PaginatedUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalCount:
          type: integer
        pageIndex:
          type: integer
        pageSize:
          type: integer
    PaginatedInvitedUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvitedUserDto'
        totalCount:
          type: integer
        pageIndex:
          type: integer
        pageSize:
          type: integer
...
    UpsertShopRequest:
      type: object
      required: [name, subdomainSlug]
      properties:
        name:
          type: string
        subdomainSlug:
          type: string
        description:
          type: string
        logoFileName:
          type: string
        bannerFileName:
          type: string
    UpdateShopActiveStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
    ShopMemberInviteRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    UpdateShopMemberRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
...